fix: toggle the mark with empth text
