+++ apps/editor/src/markdown/marks/strong.ts
@@ -2,8 +2,7 @@ import { DOMOutputSpecArray } from 'prosemirror-model';
 import { EditorCommand } from '@t/spec';
 import { cls } from '@/utils/dom';
 import Mark from '@/spec/mark';
-import { resolveSelectionPos } from '../helper/pos';
-import { createTextSelection } from '../helper/manipulation';
+import { createMarkCommand } from '../helper/mdCommand';
 
 const reStrong = /^(\*{2}|_{2}).*([\s\S]*)\1$/m;
 const strongSyntax = '**';
@@ -22,27 +21,7 @@ export class Strong extends Mark {
   }
 
   private bold(): EditorCommand {
-    return () => (state, dispatch) => {
-      const [from, to] = resolveSelectionPos(state.selection);
-      const { empty } = state.selection;
-      const slice = state.selection.content();
-      const textContent = slice.content.textBetween(0, slice.content.size, '\n');
-      let { tr } = state;
-
-      if (reStrong.test(textContent)) {
-        tr = tr.delete(to - 2, to).delete(from, from + 2);
-      } else {
-        tr = tr.insertText(strongSyntax, to).insertText(strongSyntax, from);
-        const selection = empty
-          ? createTextSelection(tr, from + 2)
-          : createTextSelection(tr, from, to + 4);
-
-        tr = tr.setSelection(selection);
-      }
-      dispatch!(tr);
-
-      return true;
-    };
+    return createMarkCommand(reStrong, strongSyntax);
   }
 
   commands() {
